+++ CotEditor/Sources/CETextSelection.h
@@ -54,7 +54,7 @@ typedef NS_ENUM(NSUInteger, CEUNFType) {
     CENFKC = 'cNkc',
     CENFKD = 'cNkd',
     CENFKCCF = 'cNcf',
-    CEAppleModifiedNFD = 'cNfa',
+    CEModifiedNFD = 'cNfm',
 };