+++ CHANGELOG.md
@@ -7,7 +7,7 @@ develop
 
 ### Additions/Changes
 
-- Add new normalization form “Apple Modified NFD” to the Unicode normalization action in Utility menu (Thanks to doraTeX!)
+- Add new normalization form “Modified NFD” (unofficial normalization form adopted by HFS+) to the Unicode normalization action in Utility menu (Thanks to doraTeX!)
     - cf. <http://tama-san.com/hfsplus_normalize/> (in Japanese)
     - It is also added to the AppleScript terms.
 - Update “JSON” syntax style: